added module to strip complexes log file
different than other output files the complexes log files are
appended run after run. Therefore, if we want to use demux on a replica
exchange log file, this may lead to errors.
The most non-invasive way to workaround this issue that I could think of
is extracting the last run within a log file. To this end I created a
separate pycomplexes module.
Up to now, this is a very dirty prototype. but I think its a good start