visualize periodic images
added a new flag for the parser.
the function is a little hacky, but I think it is the most efficient
way to handle issue: after preparation of the whole visualization
(whatever kind the user chose), I parse the resulting file for the
number of reps. based on that I add periodic images to all of them.